Rename cmd package to cli
