+++ cli/root.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"fmt"
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var rootCmd = &cobra.Command{
+var rootCommand = &cobra.Command{
 	Use:   "lfrd",
 	Short: "LFRD is a simple static website generator",
 	Long:  "LFRD (Lightweight Framework for Rapid Design) is a tool for quickly creating personal and commercial static websites",
@@ -15,12 +15,12 @@ var rootCmd = &cobra.Command{
 
 // Execute runs the root command
 func Execute() {
-	if err := rootCmd.Execute(); err != nil {
+	if err := rootCommand.Execute(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
 }
 
 func init() {
-	rootCmd.AddCommand(initCmd)
+	rootCommand.AddCommand(initCliCommand)
 }
