+++ go.mod
@@ -4,6 +4,6 @@ go 1.24.3
 
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/cobra v1.10.1 // indirect
+	github.com/spf13/cobra v1.10.1
 	github.com/spf13/pflag v1.0.9 // indirect
 )
