+++ main.go
@@ -1,7 +1,7 @@
 package main
 
-import "github.com/Rahulio2D/LFRD/cmd"
+import "github.com/Rahulio2D/LFRD/cli"
 
 func main() {
-	cmd.Execute()
+	cli.Execute()
 }
